staticfiles dirs updated

+++ art_generator/views.py
@@ -117,46 +117,51 @@ def image_generate_view(request):
             elif prompt.strip()=="":
                 return False 
         except:
-            print("image_generate_view/prompt was either not a string, or was None",prompt)     
+            print("image_generate_view/prompt was either not a string, or was None:",prompt)     
             return False 
 
         return True
     
-    # get project info
-    try:
-        #웹사이트로부터 정보를 받고
-        project_name=request.POST.get('project_name')
-        negative_prompt = request.POST.get('negative_prompt')
-        positive_prompt = request.POST.get('positive_prompt')
+    if request.method=="POST":
+        # get project info
+        try:
+            #웹사이트로부터 정보를 받고
+            project_name=request.POST.get('project_name')
+            if project.projectName!=project_name and prompt_exist(project_name):
+                project.changeName(project_name)
+                project.save()
+        except:
+            print('image_generate_view/project name was not received')
+        try:
+            negative_prompt = request.POST.get('negative_prompt')
+            positive_prompt = request.POST.get('positive_prompt')
 
-        #project 인스턴스에 정보 업데이트
-        #NEED WORK: class 함수를 왜 안씀?
-        if project.projectName!=project_name and prompt_exist(project_name):
-            project.changeName(project_name)
-        if project.negativePrompt!=negative_prompt and prompt_exist(negative_prompt):
-            project.negativePrompt=negative_prompt
-        if project.positivePrompt!=positive_prompt and prompt_exist(positive_prompt):
-            project.positivePrompt=positive_prompt
-        project.save() 
+            #project 인스턴스에 정보 업데이트
+            if project.negativePrompt!=negative_prompt and prompt_exist(negative_prompt):
+                project.negativePrompt=negative_prompt
+            if project.positivePrompt!=positive_prompt and prompt_exist(positive_prompt):
+                project.positivePrompt=positive_prompt
+            project.save() 
 
-    except:
-        #WARNING: test prompt used
-        if not prompt_exist(project.projectName):
-            project.projectName="default_project"
-        project.negativePrompt="not a fierce cat"
-        project.positivePrompt="is a cute cat"
+        except:
+            #WARNING: test prompt used
+            print('test prompt is delivered')
+            if not prompt_exist(project.projectName):
+                project.projectName="default_project"
+            project.negativePrompt="not a fierce cat"
+            project.positivePrompt="is a cute cat"
 
 
-    #이미 만들어진 이미지가 있었다면 불러오기
-    selected_images=[]
-    try:
-        GeneratedImage.objects.get(project=project)
-        generated_images=GeneratedImage.objects.filter(project=project)
-    except:
-        generated_images=[]
+        #이미 만들어진 이미지가 있었다면 불러오기
+        selected_images=[]
+        try:
+            GeneratedImage.objects.get(project=project)
+            generated_images=GeneratedImage.objects.filter(project=project)
+        except:
+            generated_images=[]
         
-    # NEED WORK: 만약 prompt가 존재한다면 clause 없음
-    if request.method == 'POST':
+        # NEED WORK: 만약 prompt가 존재한다면 clause 없음
+        
 
         #generate 버튼을 눌러 이미지를 요구했다면
         if 'generate' in request.POST:
@@ -177,13 +182,13 @@ def image_generate_view(request):
             project.save()
         
         elif 'changePrompt' in request.POST:
-            print(project.projectName,"????????????????????")
+            
             negative_prompt = request.POST.get('negative_prompt')
             positive_prompt = request.POST.get('positive_prompt')
             project.negativePrompt=negative_prompt
             project.positivePrompt=positive_prompt
             project.save() 
-            #NEED WORK: django web push를 아직 안함
+            
 
         elif 'export' in request.POST:
 
@@ -203,19 +208,22 @@ def image_generate_view(request):
         #지우기와 고르기 
         if request.POST.get('action') == 'toggle_button':
             if request.POST.get('button_type')=='remove':
+                print("!!!!!!!!!!!!REMOVE ACTIVATED")
                 token = request.POST.get('csrfmiddlewaretoken')
                 id = request.POST.get("id")#image object id 
-                GeneratedImage.objects.get(id=id).delete()
-                generated_images=GeneratedImage.objects.filter(project=project)
-                selected_images=GeneratedImage.objects.filter(project=project,selected=True)
-                redirect('art_generator:imageGenerate')
+                request.session['image_id']=id
+                #generated_images=GeneratedImage.objects.filter(project=project)
+                #selected_images=GeneratedImage.objects.filter(project=project,selected=True)
+                return redirect('art_generator:deleteImage')
 
             elif request.POST.get('button_type')=='select':
+                print("!!!!!!!!!!!!SUBMIT ACTIVATED")
                 token = request.POST.get('csrfmiddlewaretoken')
                 id = request.POST.get("id")#image object id 
+                request.session['image_id']=id
                 GeneratedImage.objects.select_image(id)
-                selected_images=GeneratedImage.objects.filter(project=project,selected=True)
-                redirect('art_generator:imageGenerate')
+                #selected_images=GeneratedImage.objects.filter(project=project,selected=True)
+                return redirect('art_generator:imageGenerate')
     #where to return
     link='imageGenerate.html'
     context={'project':project,'generated_images': generated_images, 'selected_images': selected_images}
@@ -264,8 +272,9 @@ def delete_image_view(request):
     project=Project.objects.get(id=project_id)
     image_id=request.session['image_id']
     image=GeneratedImage.objects.get(id=image_id,project=project)
-    if 'delete' in request.POST:
-        image.delete()
-        redirect('art_generator:ImageGenerate')
+    if request.method == 'POST':
+        if 'delete' in request.POST:
+            image.delete()
+            redirect('art_generator:ImageGenerate')
     context={'image':image}
     return render(request,'deleteImage.html')
